refactor: session에 userId 저장하게 변경
